Split tool into two modes.
This is to combat the issue where the `file` command line argument was
required regardless of whether it was actually needed (for example if a
control topic is provided then the files are just saved under the
timestamp).
